ordenando parte 1
