+++ Backend/src/routes/handlers/register.js
@@ -1,6 +1,4 @@
-const bcrypt = require("bcrypt");
-const { User } = require("../../db");
-const { Op } = require("sequelize");
+const registerController = require("../controllers/registerController");
 
 module.exports = async (req, res, next) => {
   try {
@@ -10,23 +8,9 @@ module.exports = async (req, res, next) => {
       return res.status(400).json({ error: "Missing data." });
     }
 
-    const saltRounds = 10;
-    const passwordHash = await bcrypt.hash(password, saltRounds);
+    const data = registerController(username, email, password);
 
-    const [newUser, alreadyExists] = await User.findOrCreate({
-      where: { [Op.or]: [{ username }, { email }] },
-      defaults: { username, email, passwordHash },
-    });
-
-    if (alreadyExists) {
-      return res.status(201).json({
-        message: "User created succesfully.",
-        accessToken: true,
-        userId: newUser.id,
-      });
-    } else {
-      return res.status(409).json({ error: "User already exist." });
-    }
+    return res.status(201).json(data);
   } catch ({ message }) {
     return res.status(500).json({ error: message });
   }
