+++ Backend/src/routes/index.js
@@ -5,9 +5,8 @@ const postEvents = require("./handlers/postEvents");
 const getAllEvents = require("./handlers/getAllEvents");
 const register = require("./handlers/register");
 const login = require("./handlers/login");
-const getEventDay = require('../routes/handlers/getEventsDay')
-const getEventsByName = require("./handlers/getEventsByName")
-
+const getEventDay = require("../routes/handlers/getEventsDay");
+const getEventsByName = require("./handlers/getEventsByName");
 
 const { Router } = require("express");
 
@@ -18,8 +17,8 @@ router.get("/months", getMonthOfYears);
 router.post("/events", postEvents);
 router.delete("/events/:id", deleteEvent);
 router.get("/events", getAllEvents);
-router.get('/eventday', getEventDay)
-router.get("/events/name",getEventsByName)
+router.get("/eventday", getEventDay);
+router.get("/events/name", getEventsByName);
 router.post("/user/register", register);
 router.post("/user/login", login);
